Fix up jtag tests
